Fix cold-start developer experience: model size, deprecation warning, quickstart comment
- README Quick Start: Qwen2.5-7B → 1.5B (fits 8GB consumer GPUs)
- from_pretrained(): torch_dtype → dtype (fixes transformers 5.2 deprecation warning)
- quickstart.py: add comment explaining empty output from random weights

+++ src/avp/connectors/huggingface.py
@@ -155,7 +155,7 @@ class HuggingFaceConnector(EngineConnector):
         has_device_map = "device_map" in model_kwargs
 
         model = transformers.AutoModelForCausalLM.from_pretrained(
-            model_id, torch_dtype=dtype, **model_kwargs
+            model_id, dtype=dtype, **model_kwargs
         )
         model.eval()
         tokenizer = transformers.AutoTokenizer.from_pretrained(model_id)
